video: tegra: dc: fake input device
Bug 855811

For smoothness and rate-limiting, it is useful to synchronize input
events to flip.  This patch adds a fake input device that emits an event
on each vsync.  This allows the Android input framework, which has
significant infrastructure to monitor input devices in particular, to be
notified of vsync events.

Because there is no input event type/code for this behavior, we use an
undocumented protocol of emitting a single MSC_RAW type event with
scan-code 1.  We also do not register ourselves for any other
buttons/keys, which should prevent our device node from being confused
with other input types.

Change-Id: Id5297d2cef93eb53737942b43b2b09b82dee5f6f
Reviewed-on: http://git-master/r/48539
Reviewed-by: Rohan Somvanshi <rsomvanshi@nvidia.com>
Tested-by: Rohan Somvanshi <rsomvanshi@nvidia.com>