arm: tegra: secureos: restore L2 AUX_CTRL during LP0 resume
In the previous code, we relied on l2x0_restart to reenable the L2.
But this interface doesn't allow the AUX_CTRL to be specified, so is
unable to pass it through to the SMC.

We now, just make the same underlying SMC call to restart the L2,
passing our saved AUX_CTRL setting.

Bug 868056

Change-Id: Ic2024bcb8df3167f00d04bba446b59696e376306
Reviewed-on: http://git-master/r/55845
Reviewed-by: Ryan Wong <ryanw@nvidia.com>
Tested-by: Ryan Wong <ryanw@nvidia.com>